requirements.txt
fastapi
deepgram-sdk
uvicorn[standard]
websockets
chromadb
langchain
langchain-community
pypdf
python-dotenv
sentence-transformers
numpy
faiss-cpu
google-generativeai

# Audio Processing Dependencies - New Providers
groq                          # Groq API for STT (Whisper Large V3)
edge-tts

# Audio processing utilities
soundfile
librosa

# Optional: Fallback TTS providers (uncomment if needed)
# edge-tts
# pyttsx3 
# google-cloud-texttospeech
